Add configuration support
There are a couple of options supported now. Users will usually want to leave
them as default but I wanted to play around with Anki's add-on config support.

Since Anki supports add-on config documentation in the `config.md` file, I have
also update the Makefile to paste the contents of that file in the appropriate
spot in the readme.